+++ src/connectors/tagService.ts
@@ -6,7 +6,7 @@ import { ARTICLE_STATE, BATCH_SIZE, MATERIALIZED_VIEW } from 'common/enums'
 import { ServerError } from 'common/errors'
 import logger from 'common/logger'
 import { BaseService } from 'connectors'
-import { GQLSearchInput } from 'definitions'
+import { GQLSearchInput, ItemData } from 'definitions'
 
 export class TagService extends BaseService {
   constructor() {
@@ -230,30 +230,63 @@ export class TagService extends BaseService {
    *********************************/
   createArticleTags = async ({
     articleIds,
-    tagIds
+    tagIds,
+    selected
   }: {
     articleIds: string[]
     tagIds: string[]
+    selected?: boolean
   }) => {
     articleIds = _.uniq(articleIds)
     tagIds = _.uniq(tagIds)
 
     const items = _.flatten(
       articleIds.map(articleId => {
-        return tagIds.map(tagId => ({ articleId, tagId }))
+        return tagIds.map(tagId => ({
+          articleId,
+          tagId,
+          ...(selected === true ? { selected } : {})
+        }))
       })
     )
     return this.baseBatchCreate(items, 'article_tag')
   }
 
   /**
+   * Update article tag.
+   */
+  putArticleTag = async ({
+    articleId,
+    tagId,
+    data
+  }: {
+    articleId: string
+    tagId: string
+    data: ItemData
+  }) =>
+    this.knex('article_tag')
+      .where({ articleId, tagId })
+      .update(data)
+      .returning('*')
+
+  /**
    * Count tags by a given tag text.
    */
-  countArticles = async (id: string) => {
+  countArticles = async ({
+    id,
+    selected
+  }: {
+    id: string
+    selected?: boolean
+  }) => {
     const result = await this.knex('article_tag')
       .join('article', 'article_id', 'article.id')
       .countDistinct('article_id')
-      .where({ tagId: id, state: ARTICLE_STATE.active })
+      .where({
+        tagId: id,
+        state: ARTICLE_STATE.active,
+        ...(selected === true ? { selected } : {})
+      })
       .first()
     return parseInt(result ? (result.count as string) : '0', 10)
   }
@@ -264,18 +297,24 @@ export class TagService extends BaseService {
   findArticleIds = async ({
     id: tagId,
     offset = 0,
-    limit = BATCH_SIZE
+    limit = BATCH_SIZE,
+    selected
   }: {
     id: string
     offset?: number
     limit?: number
     filter?: { [key: string]: any }
+    selected?: boolean
   }) => {
     const result = await this.knex
       .select('article_id')
       .from('article_tag')
       .join('article', 'article_id', 'article.id')
-      .where({ tagId, state: ARTICLE_STATE.active })
+      .where({
+        tagId,
+        state: ARTICLE_STATE.active,
+        ...(selected === true ? { selected } : {})
+      })
       .limit(limit)
       .offset(offset)
       .orderBy('article_tag.id', 'desc')