+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - 2020-02-06
+
+### Changed
+
+- Migration queue and service #669
+- Change article tag schema and rename APIs. #677
+- Alter tag API for filtering out selected or all articles under a specific tag #715
+- Update userName index field for searching #670
+- Update recommendation engine #696
+- Use `@matters/passport-likecoin` #697
+
 ## [2.6.0] - 2020-01-13
 
 ### Added